Fix failing HibernateSystemConfigStore tests.
The following 7 tests from this class fail on a clean install with `mvn
test`.

  shouldBeAbleToSetSystemProperty
  shouldGetExceptionWhenRetrievingMissingUser
  shouldBeAbleToAddUserAndSeeItByGet
  shouldBeAbleToAddUserAndSeeItInList
  shouldGetExceptionWhenRetrievingDeletedUser
  shouldNotSeeDeletedUsersInList
  shouldBeAbleToUpdateUser

These tests pass if HibernateSystemConfigStoreTest is tested in
isolation with `mvn test -Dtest=HibernateSystemConfigStoreTest`.

It seems that perhaps when multiple tests run, the same location in
memory (for the in-memory database) is used which causes the constraint
violation.

All tests pass with this change.