Prevent final date range constraints being copied to all categories.
If multiple date range categories are configured for an endpoint, the
final one added to the document will cause startDate and endDate to be
set accordingly, and this value will then be applied via
glDatePicker({...}) to every input element matching the selector
.category[data-constraint-type=date] input[type=text], i.e., every
single text input field for all date categories, as opposed to the
desired behaviour, which is to only apply the upper and lower bound
restrictions to the datepickers for the category we're adding.

This is corrected by using .category:last-child in the selector, so that
we only target the current category being added.